Добавлена возможность редактировать проект и менять его пользователей

+++ View/Controllers/ProjectController.cs
@@ -1,32 +1,20 @@
 ﻿using System.Linq;
 using System.Net;
+using System.Runtime.InteropServices.ComTypes;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Internal;
 using Timetracker.Entities.Classes;
 using Timetracker.Entities.Models;
 
 namespace View.Controllers
 {
-    public class ProjectsView
-    {
-        public Project[] SignedProjects { get; set; }
-
-        public Project[] NotSignedProjects { get; set; }
-    }
-
     public class AddProjectView
     {
         public Project Project { get; set; }
-
-        public string[] Users { get; set; }
-    }
-
-    public class InviteUsersView
-    {
-        public int ProjectId { get; set; }
 
         public string[] Users { get; set; }
     }
@@ -37,32 +25,38 @@ namespace View.Controllers
     public class ProjectController : ControllerBase
     {
         private readonly TimetrackerContext _dbContext;
+        private readonly JsonSerializerOptions _jsonOptions;
 
         public ProjectController(TimetrackerContext authorizedUsersRepository)
         {
             _dbContext = authorizedUsersRepository;
+
+            _jsonOptions = new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            };
         }
 
-        public async Task<IActionResult> GetProjects()
+        public async Task<IActionResult> GetAll()
         {
             var user = await _dbContext.GetUser(User.Identity.Name);
 
             var userProjects = _dbContext.AuthorizedUsers.AsNoTracking()
-                .Where(x => x.UserId == user.Id);
+                .Where(x => x.UserId == user.Id)
+                .Include(x => x.Project)
+                .Select(x => new { x.IsSigned, x.Project });
 
             var projects = await userProjects
                 .Where(x => x.IsSigned)
-                .Include(x => x.Project)
                 .Select(x => x.Project)
                 .ToArrayAsync();
 
             var notSignedProjects = await userProjects
                 .Where(x => !x.IsSigned)
-                .Include(x => x.Project)
                 .Select(x => x.Project)
                 .ToArrayAsync();
 
-            var projectsView = new ProjectsView
+            var projectsView = new
             {
                 SignedProjects = projects,
                 NotSignedProjects = notSignedProjects
@@ -86,9 +80,10 @@ namespace View.Controllers
                 .FirstOrDefaultAsync(x => x.Id == id);
 
             var users = await _dbContext.AuthorizedUsers
+                .AsNoTracking()
                 .Where(x => x.ProjectId == id)
                 .Include(x => x.User)
-                .Select(x => new { x.User.Login })
+                .Select(x => x.User.Login)
                 .ToArrayAsync();
 
             var response = new {
@@ -97,17 +92,17 @@ namespace View.Controllers
                 users = users
             };
 
-            var jsonOptions = new JsonSerializerOptions
-            {
-                PropertyNameCaseInsensitive = true
-            };
-
-            return new JsonResult(response, jsonOptions);
+            return new JsonResult(response, _jsonOptions);
         }
 
         public async Task<IActionResult> AddProject([FromBody] AddProjectView project)
         {
             var user = await _dbContext.GetUser(User.Identity.Name);
+
+            if (_dbContext.Projects.Any(x => x.Title == project.Project.Title))
+            {
+                return BadRequest();
+            }
 
             await _dbContext.AddAsync(project.Project);
             await _dbContext.AddAsync(new AuthorizedUser
@@ -121,6 +116,9 @@ namespace View.Controllers
             foreach (var userName in project.Users)
             {
                 var userInviting = await _dbContext.GetUser(userName);
+
+                if (userInviting == null)
+                    continue;
 
                 await _dbContext.AddAsync(new AuthorizedUser
                 {
@@ -136,6 +134,45 @@ namespace View.Controllers
             return Ok();
         }
 
+        public async Task<IActionResult> Update([FromBody] AddProjectView updatedProject)
+        {
+            _dbContext.Projects.Update(updatedProject.Project);
+
+            var user = await _dbContext.GetUser(User.Identity.Name);
+
+            var authorizedUsers = await _dbContext.AuthorizedUsers.AsNoTracking()
+                .Where(x => x.ProjectId == updatedProject.Project.Id && x.UserId != user.Id)
+                .Include(x => x.User)
+                .Where(x => !updatedProject.Users.Any(y => x.User.Login == y ) )
+                .ToArrayAsync();
+
+            if (authorizedUsers.Any())
+                _dbContext.RemoveRange(authorizedUsers);
+
+            foreach (var userName in updatedProject.Users)
+            {
+                if (userName == User.Identity.Name)
+                    continue;
+
+                var userInviting = await _dbContext.GetUser(userName);
+
+                if (userInviting == null)
+                    continue;
+
+                await _dbContext.AddAsync(new AuthorizedUser
+                {
+                    IsSigned = false,
+                    Project = updatedProject.Project,
+                    RightId = 1,
+                    UserId = userInviting.Id
+                });
+            }
+
+            _dbContext.SaveChanges();
+
+            return Ok();
+        }
+
         public async Task<IActionResult> Remove(int? id)
         {
             if (!id.HasValue)
@@ -146,12 +183,22 @@ namespace View.Controllers
                 });
             }
 
+            var project = await _dbContext.Projects.FirstOrDefaultAsync(x => x.Id == id);
+
+            if (project == null)
+            {
+                return new JsonResult(new
+                {
+                    text = "Не существует проекта с таким Id",
+                    status = HttpStatusCode.BadRequest
+                });
+            }
+
             var authorizedUsers = _dbContext.AuthorizedUsers
                 .Where(x => x.ProjectId == id).ToList();
             
             _dbContext.RemoveRange(authorizedUsers);
 
-            var project = await _dbContext.Projects.FirstOrDefaultAsync(x => x.Id == id);
             _dbContext.Remove(project);
 
             await _dbContext.SaveChangesAsync();
@@ -160,29 +207,6 @@ namespace View.Controllers
             {
                 status = HttpStatusCode.OK
             });
-        }
-
-        public async Task<IActionResult> InviteUsers([FromBody] InviteUsersView view)
-        {
-            var project = await _dbContext.Projects.AsNoTracking()
-                .FirstOrDefaultAsync(x => x.Id == view.ProjectId);
-
-            foreach (var userName in view.Users)
-            {
-                var userInviting = await _dbContext.GetUser(userName);
-
-                await _dbContext.AddAsync(new AuthorizedUser
-                {
-                    IsSigned = false,
-                    Project = project,
-                    RightId = 1,
-                    UserId = userInviting.Id
-                });
-            }
-
-            await _dbContext.SaveChangesAsync();
-
-            return Ok();
         }
     }
 }
